+++ crates/gpui/src/platform/mac/open_type.rs
@@ -1,6 +1,6 @@
 #![allow(unused, non_upper_case_globals)]
 
-use crate::FontFeatures;
+use crate::{FontFallbacks, FontFeatures};
 use cocoa::appkit::CGFloat;
 use core_foundation::{
     array::{
@@ -15,17 +15,74 @@ use core_foundation::{
 };
 use core_graphics::{display::CFDictionary, geometry::CGAffineTransform};
 use core_text::{
-    font::{CTFont, CTFontRef},
+    font::{cascade_list_for_languages, CTFont, CTFontRef},
     font_descriptor::{
-        kCTFontFeatureSettingsAttribute, CTFontDescriptor, CTFontDescriptorCopyAttributes,
-        CTFontDescriptorCreateCopyWithFeature, CTFontDescriptorCreateWithAttributes,
+        kCTFontCascadeListAttribute, kCTFontFeatureSettingsAttribute, CTFontDescriptor,
+        CTFontDescriptorCopyAttributes, CTFontDescriptorCreateCopyWithFeature,
+        CTFontDescriptorCreateWithAttributes, CTFontDescriptorCreateWithNameAndSize,
         CTFontDescriptorRef,
     },
 };
-use font_kit::font::Font;
+use font_kit::font::Font as FontKitFont;
 use std::ptr;
 
-pub fn generate_feature_array(features: &FontFeatures) -> CFMutableArrayRef {
+pub fn apply_features_and_fallbacks(
+    font: &mut FontKitFont,
+    features: &FontFeatures,
+    fallbacks: &FontFallbacks,
+    pref_langs: &CFArray<CFString>,
+) -> anyhow::Result<()> {
+    unsafe {
+        let fallback_array = CFArrayCreateMutable(kCFAllocatorDefault, 0, &kCFTypeArrayCallBacks);
+
+        for user_fallback in fallbacks.fallback_list() {
+            let name = CFString::from(user_fallback.as_str());
+            let fallback_desc =
+                CTFontDescriptorCreateWithNameAndSize(name.as_concrete_TypeRef(), 0.0);
+            CFArrayAppendValue(fallback_array, fallback_desc as _);
+            CFRelease(fallback_desc as _);
+        }
+
+        cascade_list_for_languages(&font.native_font(), pref_langs)
+            .into_iter()
+            .filter(|desc| desc.font_path().is_some())
+            .for_each(|user_fallback_desc| {
+                CFArrayAppendValue(
+                    fallback_array,
+                    user_fallback_desc.as_concrete_TypeRef() as _,
+                );
+            });
+
+        let feature_array = generate_feature_array(features);
+        let keys = [kCTFontFeatureSettingsAttribute, kCTFontCascadeListAttribute];
+        let values = [feature_array, fallback_array];
+        let attrs = CFDictionaryCreate(
+            kCFAllocatorDefault,
+            keys.as_ptr() as _,
+            values.as_ptr() as _,
+            2,
+            &kCFTypeDictionaryKeyCallBacks,
+            &kCFTypeDictionaryValueCallBacks,
+        );
+        CFRelease(feature_array as *const _ as _);
+        CFRelease(fallback_array as *const _ as _);
+        let new_descriptor = CTFontDescriptorCreateWithAttributes(attrs);
+        CFRelease(attrs as _);
+        let new_descriptor = CTFontDescriptor::wrap_under_create_rule(new_descriptor);
+        let new_font = CTFontCreateCopyWithAttributes(
+            font.native_font().as_concrete_TypeRef(),
+            0.0,
+            std::ptr::null(),
+            new_descriptor.as_concrete_TypeRef(),
+        );
+        let new_font = CTFont::wrap_under_create_rule(new_font);
+        *font = font_kit::font::Font::from_native_font(&new_font);
+
+        Ok(())
+    }
+}
+
+fn generate_feature_array(features: &FontFeatures) -> CFMutableArrayRef {
     unsafe {
         let mut feature_array =
             CFArrayCreateMutable(kCFAllocatorDefault, 0, &kCFTypeArrayCallBacks);
